requirements.txt
tensorflow==2.15.0
streamlit==1.38.0
numpy==1.24.3
pillow
